+++ src/CakeBoss.Host/Aliases/ScheduleAliases.cs
@@ -17,7 +17,36 @@ namespace CakeBoss.Host
     public static class ScheduleAliases
     {
         /// <summary>
-        /// Starts a Nuget server using the specified information.
+        /// Starts the scheduled task manager
+        /// </summary>
+        /// <param name="context">The context.</param>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Schedule")]
+        public static void StartSchedules(this ICakeContext context)
+        {
+            TaskRegistry instance = context.GetContainer().GetInstance<TaskRegistry>();
+            TaskManager.Initialize(instance);
+
+            context.Log.Information("Scheduled tasks started.");
+        }
+        
+        /// <summary>
+        /// Stops the scheduled task manager
+        /// </summary>
+        /// <param name="context">The context.</param>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Schedule")]
+        public static void StopSchedules(this ICakeContext context)
+        {
+            TaskManager.Stop();
+
+            context.Log.Information("Scheduled tasks stopped.");
+        }
+
+
+
+        /// <summary>
+        /// Adds a scheduled task to the task manager
         /// </summary>
         /// <param name="context">The context.</param>
         /// <param name="target">The task you want to schedule.</param>
@@ -25,24 +54,10 @@ namespace CakeBoss.Host
         [CakeAliasCategory("Schedule")]
         public static Schedule ScheduleTask(this ICakeContext context, string target)
         {
-            TaskRegistry instance = context.GetContainer().GetInstance<TaskRegistry>();
+            context.Log.Information("Scheduling task '{0}'.", target);
 
+            TaskRegistry instance = context.GetContainer().GetInstance<TaskRegistry>();
             return instance.Schedule(target);
-        }
-
-        /// <summary>
-        /// Starts a Nuget server using the specified information.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <param name="target">The task you want to shedule.</param>
-        [CakeMethodAlias]
-        [CakeAliasCategory("Schedule")]
-        public static void StartSchedules(this ICakeContext context)
-        {
-            context.Log.Information("Starting scheduled tasks.");
-
-            TaskRegistry instance = context.GetContainer().GetInstance<TaskRegistry>();
-            TaskManager.Initialize(instance);
         }
     }
 }